// src/api/generated/api.ts
export interface AdminContactsIdPatchRequest {
    'status'?: string;
}
export interface AdminJobsIdPatchRequest {
    'title'?: string;
    'department'?: string;
    'location'?: string;
    'description'?: string;
    'salary'?: string;
}
export interface AdminJobsIdStatusPatchRequest {
    'isActive'?: boolean;
}
export interface AdminLoginPost200Response {
    'success'?: boolean;
    'token'?: string;
}
export interface AdminLoginPostRequest {
    'email': string;
    'password': string;
}
// ...
export interface JobsPostRequest {
    'title': string;
    'department': string;
    'location': string;
    'type'?: string;
    'salary'?: string;
    'description': string;
    'responsibilities'?: Array<string>;
    'requirements'?: Array<string>;
    'isActive'?: boolean;
}
// ...
export class AdminApi extends BaseAPI {
    /**
     * 
     * @summary Update application status
     * @param {string} appId Application ID
     * @param {AdminContactsIdPatchRequest} adminContactsIdPatchRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public adminApplicationsAppIdStatusPatch(appId: string, adminContactsIdPatchRequest: AdminContactsIdPatchRequest, options?: RawAxiosRequestConfig) {
        return AdminApiFp(this.configuration).adminApplicationsAppIdStatusPatch(appId, adminContactsIdPatchRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get all general job applications
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public adminApplicationsGet(options?: RawAxiosRequestConfig) {
        return AdminApiFp(this.configuration).adminApplicationsGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get all contact submissions
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public adminContactsGet(options?: RawAxiosRequestConfig) {
        return AdminApiFp(this.configuration).adminContactsGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get a contact by ID
     * @param {string} id Contact ID
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public adminContactsIdGet(id: string, options?: RawAxiosRequestConfig) {
        return AdminApiFp(this.configuration).adminContactsIdGet(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Update contact status
     * @param {string} id Contact ID
     * @param {AdminContactsIdPatchRequest} adminContactsIdPatchRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public adminContactsIdPatch(id: string, adminContactsIdPatchRequest: AdminContactsIdPatchRequest, options?: RawAxiosRequestConfig) {
        return AdminApiFp(this.configuration).adminContactsIdPatch(id, adminContactsIdPatchRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get all job listings (including inactive)
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public adminJobsGet(options?: RawAxiosRequestConfig) {
        return AdminApiFp(this.configuration).adminJobsGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Delete a job listing
     * @param {string} id Job ID
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public adminJobsIdDelete(id: string, options?: RawAxiosRequestConfig) {
        return AdminApiFp(this.configuration).adminJobsIdDelete(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Update job listing details
     * @param {string} id Job ID
     * @param {AdminJobsIdPatchRequest} adminJobsIdPatchRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public adminJobsIdPatch(id: string, adminJobsIdPatchRequest: AdminJobsIdPatchRequest, options?: RawAxiosRequestConfig) {
        return AdminApiFp(this.configuration).adminJobsIdPatch(id, adminJobsIdPatchRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Toggle job listing status (active/inactive)
     * @param {string} id Job ID
     * @param {AdminJobsIdStatusPatchRequest} adminJobsIdStatusPatchRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public adminJobsIdStatusPatch(id: string, adminJobsIdStatusPatchRequest: AdminJobsIdStatusPatchRequest, options?: RawAxiosRequestConfig) {
        return AdminApiFp(this.configuration).adminJobsIdStatusPatch(id, adminJobsIdStatusPatchRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Authenticate admin user
     * @param {AdminLoginPostRequest} adminLoginPostRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public adminLoginPost(adminLoginPostRequest: AdminLoginPostRequest, options?: RawAxiosRequestConfig) {
        return AdminApiFp(this.configuration).adminLoginPost(adminLoginPostRequest, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class JobsApi extends BaseAPI {
    /**
     * 
     * @summary Get all active job listings
     * @param {boolean} [showAll] Show all jobs (Admin only)
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public jobsGet(showAll?: boolean, options?: RawAxiosRequestConfig) {
        return JobsApiFp(this.configuration).jobsGet(showAll, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Create a new job listing (Admin)
     * @param {JobsPostRequest} jobsPostRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public jobsPost(jobsPostRequest: JobsPostRequest, options?: RawAxiosRequestConfig) {
        return JobsApiFp(this.configuration).jobsPost(jobsPostRequest, options).then((request) => request(this.axios, this.basePath));
    }
}